backend/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
pydantic==2.7.4
langchain==0.1.16
langchain-google-genai==1.0.1
google-generativeai==0.4.0
python-dotenv==1.0.0
